\name{DiagTest3Grp-internal}
\title{Internal DiagTest3Grp functions}
\alias{Bootstrap.Youden.Cutoff.1run}
%\alias{BW.ref}
%\alias{CI.normal}
%\alias{empirical.cdf}
%\alias{FisherZ}
%\alias{FisherZ.Var}
%\alias{KernelSmoothing.cdf}
\alias{negloglik.boxcox.3Grp}
%\alias{NonParametric.VUS}
%alias{NonParametric.VUS.var}
%\alias{Normal.VUS}
%\alias{Normal.Youden.Cutoff.CI}
\alias{PartialDeriv.optimalCutoff}
\alias{PartialDeriv.Youden}
%\alias{SampleSize.VUS}
%\alias{SampleSize.Youden3Grp}
%\alias{var.mu}
%\alias{var.sigma}
\alias{Var.t.minus}
\alias{Var.Youden}
\alias{Youden3Grp.boxcox.fit}
\alias{Youden3Grp.BoxCox.lambda}
\alias{Youden3Grp.Normal}
%\alias{Youden3Grp.Variance.Bootstrap}
\description{Internal DiagTest3Grp functions}
\details{These are not to be called by the users or are undocumented}
\keyword{internal}
